Migrating BranchOffice and Operator

+++ FishBusiness/Controllers/IMerchantRecieptsController.cs
@@ -111,7 +111,7 @@ namespace FishBusiness.Controllers
                 string[] Fishes = FishesCookie.Split(",").Select(c => Convert.ToString(c)).ToArray();
                 string[] Productions = ProductionTypesCookie.Split(",").Select(c => Convert.ToString(c)).ToArray();
 
-                int[] qtys = qtysCookie.Split(",").Select(c => Convert.ToInt32(c)).ToArray();
+                double[] qtys = qtysCookie.Split(",").Select(c => Convert.ToDouble(c)).ToArray();
                 decimal[] unitPrices = unitpricesCookie.Split(",").Select(c => Convert.ToDecimal(c)).ToArray();
 
                 IMerchantReciept imerchantReciept;
@@ -276,7 +276,7 @@ namespace FishBusiness.Controllers
             Fish fish = _context.Fishes.Find(item.FishID);
 
             ProductionType production = _context.ProductionTypes.Find(item.ProductionTypeID);
-            var res = new { productionName = production.ProductionName, fishName = fish.FishName, qty = item.Qty, unitPrice = item.UnitPrice, total = item.Qty * item.UnitPrice };
+            var res = new { productionName = production.ProductionName, fishName = fish.FishName, qty = item.Qty, unitPrice = item.UnitPrice, total = (decimal) item.Qty * item.UnitPrice };
 
             return Json(res);
