+++ FishBusiness/Models/MerchantRecieptItem.cs
@@ -14,7 +14,7 @@ namespace FishBusiness.Models
 
         [Display(Name = "الكمية")]
         [Required(ErrorMessage = "برجاء ادخال الكمية")]
-        public int Qty { get; set; }
+        public double Qty { get; set; }
 
         [Display(Name = "سعر الوحدة")]
         [Required(ErrorMessage = "برجاء ادخال سعر الوحدة")]
